chore: Move sorted tracks function over
